moving the getThem function from helpers.js to utlis.js, added primary color to buttons and deleted the detectTheme function
